Fixed the method to calculate the remaining games.

+++ assignment4/board.h
@@ -66,5 +66,5 @@ class Board {
 
     void clean();
 
-    int calculate();
+    long calculate();
 };
